Added takeDamage and checkIfDead methods to Gladiator class.
I added two new methods to the Gladiator class. The first, takeDamage,
allows for the player to be attacked. I created this method, rather than
using a setter, so that no one could add to a players health without
following my set rule. I also added a checkIfDead method that, if the
player's health is less than or equal to zero it will set a Boolean
variable called isDead to true.